Fix compilation under Clang
* iB4e code used (int) to store pointers, which fails on 64-bit systems.
  GCC knows how to fix this, but Clang doesn't.

+++ iB4e/stack.cc
@@ -76,7 +76,7 @@ int sortcompare(void *a, void *b)
   face1 = (Face *) a; 
   face2 = (Face *) b;
 
-  int compare1, compare2;
+  size_t compare1, compare2;
 
   if(face1->numvertices != face2->numvertices) {
     cerr << "Number of vertices mismatch in sortcompare!!";
@@ -89,8 +89,8 @@ int sortcompare(void *a, void *b)
 
   for(int i = 0; i < face1->numvertices; i++) {
 
-    compare1 = (int)(face1->vertices[i]);
-    compare2 = (int)(face2->vertices[i]);
+    compare1 = (size_t)(face1->vertices[i]);
+    compare2 = (size_t)(face2->vertices[i]);
 
     #ifdef DEBUG
     cout << "\n" << compare1 << " vs " << compare2;
